+++ .gitignore
@@ -28,3 +28,5 @@ dist-ssr
 
 # expo dir
 .expo/
+
+cmd-helper.ipynb
